updating with new inputs

+++ 2018/day1/part1.py
@@ -1,6 +1,6 @@
 count = 0
 with open('day1.txt') as f:
   for line in f:
-    frequency = float(line.strip())
+    frequency = int(line.strip())
     count += frequency
 print(count)
