+++ 2018/day1/part2.py
@@ -4,7 +4,7 @@ boats = True
 while boats:
   with open('day1.txt') as f:
     for line in f:
-      delta = float(line.strip())
+      delta = int(line.strip())
       frequency += delta 
       if frequency in frequency_list:
         print("Duplicate Frequency Found:", frequency)
